Minor: Avulse folder deletion, checklist form refactoring, folder creation for multiple checklist model organization
