--- src/app/checklists/checklist.component.html
@@ -1,3 +1,0 @@
-<div clas="container">
-    <h1>CHECKLIST</h1>
-</div>
