--- src/app/checklists/checklist.component.ts
@@ -1,10 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-    moduleId: module.id,
-    selector: 'checklist',
-    templateUrl: 'checklist.component.html'
-})
-export class ChecklistComponent{
-
-}
