--- src/app/checklists/checklist.module.ts
@@ -1,13 +1,0 @@
-import { NgModule} from '@angular/core';
-import {ChecklistComponent} from './checklist.component';
-
-@NgModule({
-    declarations:[
-        ChecklistComponent
-    ],
-    exports:[
-        ChecklistComponent
-    ]
-
-})
-export class ChecklistModule{}
