+++ src/app/models/index.ts
@@ -1,10 +1,10 @@
 export * from './user';
 export * from './scheduling.model';
-export * from './checklistForms';
+export * from './checklist/checklistForms';
 export * from './council-group';
 export * from './alert.model';
 export * from './search.model';
 export * from './state.model';
-export * from './checklist.model';
+export * from './checklist/checklist.model';
 export * from './search.model';
 export * from './posts/post.model';
